app/queue.py
# Queue and it's operation


class Node:
    def __init__(self, value) -> None:
        self.value = value
        self.next = None


class Queue:
    def __init__(self, value) -> None:
        new_node = Node(value)
        self.first = new_node
        self.last = new_node
        self.length = 1

    def print_queue(self):
        result = list()
        temp = self.first
        while temp:
            result.append(temp.value)
            temp = temp.next
        return "->".join(str(i) for i in result)

    def enqueue(self, value):  # just like `append` in linked list O(1)
        new_node = Node(value)
        if self.length == 0:
            self.first = new_node
            self.last = new_node
        else:
            self.last.next = new_node
            self.last = new_node
        self.length += 1

    def dequeue(self):  # just like `pop_first` in linked list O(1)
        if self.length == 0:
            return None

        temp = self.first
        if self.length == 1:
            self.first = None
            self.last = None
        else:
            self.first = temp.next
            temp.next = None
        self.length -= 1
        return temp
